addressed pr comments: added AuthenticationMethod enum, better authorizer error messages, separate folder for tvm handler
