+++ examples/nodejs/token-vending-machine/README.md
@@ -27,7 +27,7 @@ Before you deploy the Token Vending Machine, you will need to configure the scop
 
 You will also need to specify the authentication method for the Token Vending Machine's API Gateway Endpoint. You may choose to leave it open, allowing the API Gateway URL to be publicly accessible, or you may choose to use Lambda Authorizer or Amazon Cognito as an authenticator. Basic examples for each of these methods is provided in the [Lambda Authorizer handler](./lambda/authorizer/authorizer.ts) and [CDK definition](./infrastructure/lib/token-vending-machine-stack.ts) files.
 
-These three required configuration variables live in the [lambda/config.ts](./lambda/config.ts) file.
+These three required configuration variables live in the [lambda/config.ts](./lambda/token-vending-machine/config.ts) file.
 
 ## Deploying the Token Vending Machine App
