+++ examples/nodejs/token-vending-machine/lambda/token-vending-machine/.npmignore
