+++ examples/nodejs/token-vending-machine/lambda/token-vending-machine/.prettierrc.json
