+++ examples/nodejs/token-vending-machine/lambda/token-vending-machine/jest.config.js
